feat: remove the redundant database view files
