+++ frontend/appflowy_flutter/lib/plugins/database_view/widgets/database_view_widget.dart
@@ -3,10 +3,12 @@ import 'package:appflowy/workspace/application/view/view_ext.dart';
 import 'package:appflowy/workspace/application/view/view_listener.dart';
 import 'package:appflowy_backend/protobuf/flowy-folder2/view.pb.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 
 class DatabaseViewWidget extends StatefulWidget {
-  const DatabaseViewWidget({required this.view, super.key});
+  const DatabaseViewWidget({
+    super.key,
+    required this.view,
+  });
 
   final ViewPB view;
 
@@ -16,38 +18,21 @@ class DatabaseViewWidget extends StatefulWidget {
 
 class _DatabaseViewWidgetState extends State<DatabaseViewWidget> {
   /// Listens to the view updates.
-  late ViewListener _listener;
+  late final ViewListener _listener;
 
   /// Notifies the view layout type changes. When the layout type changes,
   /// the widget of the view will be updated.
-  late ViewLayoutTypeChangeNotifier _layoutTypeChangeNotifier;
+  late final ValueNotifier<ViewLayoutPB> _layoutTypeChangeNotifier;
 
   /// The view will be updated by the [ViewListener].
   late ViewPB view;
 
   @override
   void initState() {
-    view = widget.view;
-    _listener = ViewListener(view: widget.view);
-    _listener.start(
-      onViewUpdated: (result) {
-        result.fold(
-          (updatedView) {
-            if (mounted) {
-              view = updatedView;
-              _layoutTypeChangeNotifier.setLayoutType(view.layout);
-            }
-          },
-          (r) => null,
-        );
-      },
-    );
-
-    _layoutTypeChangeNotifier = ViewLayoutTypeChangeNotifier(
-      layoutType: widget.view.layout,
-    );
-
     super.initState();
+
+    view = widget.view;
+    _listenOnViewUpdated();
   }
 
   @override
@@ -59,30 +44,32 @@ class _DatabaseViewWidgetState extends State<DatabaseViewWidget> {
 
   @override
   Widget build(BuildContext context) {
-    return ChangeNotifierProvider<ViewLayoutTypeChangeNotifier>.value(
-      value: _layoutTypeChangeNotifier,
-      child: Consumer<ViewLayoutTypeChangeNotifier>(
-        builder: (context, notifier, _) {
-          return makePlugin(pluginType: view.pluginType, data: view)
-              .widgetBuilder
-              .buildWidget();
-        },
-      ),
+    return ValueListenableBuilder<ViewLayoutPB>(
+      valueListenable: _layoutTypeChangeNotifier,
+      builder: (_, __, ___) {
+        return makePlugin(pluginType: view.pluginType, data: view)
+            .widgetBuilder
+            .buildWidget();
+      },
     );
   }
-}
 
-class ViewLayoutTypeChangeNotifier extends ChangeNotifier {
-  ViewLayoutPB _layoutType;
+  void _listenOnViewUpdated() {
+    _listener = ViewListener(view: widget.view)
+      ..start(
+        onViewUpdated: (result) {
+          result.fold(
+            (updatedView) {
+              if (mounted) {
+                view = updatedView;
+                _layoutTypeChangeNotifier.value = view.layout;
+              }
+            },
+            (r) => null,
+          );
+        },
+      );
 
-  ViewLayoutTypeChangeNotifier({
-    required ViewLayoutPB layoutType,
-  }) : _layoutType = layoutType;
-
-  ViewLayoutPB get layoutType => _layoutType;
-
-  void setLayoutType(ViewLayoutPB layoutType) {
-    _layoutType = layoutType;
-    notifyListeners();
+    _layoutTypeChangeNotifier = ValueNotifier(widget.view.layout);
   }
 }
