+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/base/insert_page_command.dart
@@ -1,19 +1,10 @@
 import 'package:appflowy/generated/locale_keys.g.dart';
 import 'package:appflowy/plugins/database_view/application/database_view_service.dart';
-import 'package:appflowy/plugins/document/presentation/editor_plugins/board/board_node_widget.dart';
-import 'package:appflowy/plugins/document/presentation/editor_plugins/calendar/calendar_node_widget.dart';
-import 'package:appflowy/plugins/document/presentation/editor_plugins/grid/grid_node_widget.dart';
+import 'package:appflowy/plugins/document/presentation/editor_plugins/plugins.dart';
 import 'package:appflowy/workspace/application/view/view_service.dart';
 import 'package:appflowy_backend/protobuf/flowy-folder2/view.pb.dart';
 import 'package:appflowy_editor/appflowy_editor.dart';
 import 'package:easy_localization/easy_localization.dart';
-
-class DatabaseBlockKeys {
-  const DatabaseBlockKeys._();
-
-  static const String parentID = 'parent_id';
-  static const String viewID = 'view_id';
-}
 
 extension InsertDatabase on EditorState {
   Future<void> insertInlinePage(String parentViewId, ViewPB childView) async {
@@ -104,11 +95,11 @@ extension InsertDatabase on EditorState {
   String _convertPageType(ViewPB viewPB) {
     switch (viewPB.layout) {
       case ViewLayoutPB.Grid:
-        return GridBlockKeys.type;
+        return DatabaseBlockKeys.gridType;
       case ViewLayoutPB.Board:
-        return BoardBlockKeys.type;
+        return DatabaseBlockKeys.boardType;
       case ViewLayoutPB.Calendar:
-        return CalendarBlockKeys.type;
+        return DatabaseBlockKeys.calendarType;
       default:
         throw Exception('Unknown layout type');
     }
