--- frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/calendar/calendar_node_widget.dart
@@ -1,5 +1,0 @@
-class CalendarBlockKeys {
-  const CalendarBlockKeys._();
-
-  static const String type = 'calendar';
-}
