+++ frontend/appflowy_flutter/lib/plugins/document/presentation/editor_plugins/plugins.dart
@@ -1,6 +1,3 @@
-export 'board/board_node_widget.dart';
-export 'board/board_menu_item.dart';
-export 'board/board_view_menu_item.dart';
 export 'callout/callout_block_component.dart';
 export 'code_block/code_block_component.dart';
 export 'code_block/code_block_shortcut_event.dart';
@@ -9,9 +6,9 @@ export 'cover/cover_node_widget.dart';
 export 'cover/cover_image_picker.dart';
 export 'emoji_picker/emoji_menu_item.dart';
 export 'extensions/flowy_tint_extension.dart';
-export 'grid/grid_menu_item.dart';
-export 'grid/grid_node_widget.dart';
-export 'grid/grid_view_menu_item.dart';
+export 'database/inline_database_menu_item.dart';
+export 'database/database_view_block_component.dart';
+export 'database/inline_database_menu_item.dart';
 export 'math_equation/math_equation_block_component.dart';
 export 'openai/widgets/auto_completion_node_widget.dart';
 export 'openai/widgets/smart_edit_node_widget.dart';
